+++ editorsnotes_app/src/views/edit_scan_list.js
